+++ relay-event-normalization/src/eap/ai.rs
@@ -206,7 +206,21 @@ fn normalize_ai_costs(attributes: &mut Attributes, model_metadata: Option<&Model
 
     // Overwrite all values, the attributes should reflect the values we used to calculate the total.
     attributes.insert(GEN_AI__COST__INPUT_TOKENS, costs.input);
+    attributes.insert(
+        GEN_AI__COST__CACHE_READ__INPUT_TOKENS,
+        costs.cache_read_input,
+    );
+    attributes.insert(
+        GEN_AI__COST__CACHE_CREATION__INPUT_TOKENS,
+        costs.cache_creation_input,
+    );
+
     attributes.insert(GEN_AI__COST__OUTPUT_TOKENS, costs.output);
+    attributes.insert(
+        GEN_AI__COST__REASONING__OUTPUT_TOKENS,
+        costs.reasoning_output,
+    );
+
     attributes.insert(GEN_AI__COST__TOTAL_TOKENS, costs.total());
 }
 
@@ -305,6 +319,14 @@ mod tests {
 
         assert_annotated_snapshot!(attributes, @r#"
         {
+          "gen_ai.cost.cache_creation.input_tokens": {
+            "type": "double",
+            "value": 0.0
+          },
+          "gen_ai.cost.cache_read.input_tokens": {
+            "type": "double",
+            "value": 20.0
+          },
           "gen_ai.cost.input_tokens": {
             "type": "double",
             "value": 25.0
@@ -312,6 +334,10 @@ mod tests {
           "gen_ai.cost.output_tokens": {
             "type": "double",
             "value": 50.0
+          },
+          "gen_ai.cost.reasoning.output_tokens": {
+            "type": "double",
+            "value": 30.0
           },
           "gen_ai.cost.total_tokens": {
             "type": "double",
@@ -374,6 +400,14 @@ mod tests {
 
         assert_annotated_snapshot!(attributes, @r#"
         {
+          "gen_ai.cost.cache_creation.input_tokens": {
+            "type": "double",
+            "value": 0.0
+          },
+          "gen_ai.cost.cache_read.input_tokens": {
+            "type": "double",
+            "value": 0.0
+          },
           "gen_ai.cost.input_tokens": {
             "type": "double",
             "value": 90.0
@@ -381,6 +415,10 @@ mod tests {
           "gen_ai.cost.output_tokens": {
             "type": "double",
             "value": 100.0
+          },
+          "gen_ai.cost.reasoning.output_tokens": {
+            "type": "double",
+            "value": 0.0
           },
           "gen_ai.cost.total_tokens": {
             "type": "double",
@@ -483,6 +521,14 @@ mod tests {
 
         assert_annotated_snapshot!(attributes, @r#"
         {
+          "gen_ai.cost.cache_creation.input_tokens": {
+            "type": "double",
+            "value": 0.0
+          },
+          "gen_ai.cost.cache_read.input_tokens": {
+            "type": "double",
+            "value": 0.0
+          },
           "gen_ai.cost.input_tokens": {
             "type": "double",
             "value": 90.0
@@ -490,6 +536,10 @@ mod tests {
           "gen_ai.cost.output_tokens": {
             "type": "double",
             "value": 100.0
+          },
+          "gen_ai.cost.reasoning.output_tokens": {
+            "type": "double",
+            "value": 0.0
           },
           "gen_ai.cost.total_tokens": {
             "type": "double",
@@ -552,6 +602,14 @@ mod tests {
 
         assert_annotated_snapshot!(attributes, @r#"
         {
+          "gen_ai.cost.cache_creation.input_tokens": {
+            "type": "double",
+            "value": 0.0
+          },
+          "gen_ai.cost.cache_read.input_tokens": {
+            "type": "double",
+            "value": 0.0
+          },
           "gen_ai.cost.input_tokens": {
             "type": "double",
             "value": 90.0
@@ -559,6 +617,10 @@ mod tests {
           "gen_ai.cost.output_tokens": {
             "type": "double",
             "value": 100.0
+          },
+          "gen_ai.cost.reasoning.output_tokens": {
+            "type": "double",
+            "value": 0.0
           },
           "gen_ai.cost.total_tokens": {
             "type": "double",
@@ -683,6 +745,14 @@ mod tests {
             "type": "integer",
             "value": 100000
           },
+          "gen_ai.cost.cache_creation.input_tokens": {
+            "type": "double",
+            "value": 0.0
+          },
+          "gen_ai.cost.cache_read.input_tokens": {
+            "type": "double",
+            "value": 0.0
+          },
           "gen_ai.cost.input_tokens": {
             "type": "double",
             "value": 300.0
@@ -690,6 +760,10 @@ mod tests {
           "gen_ai.cost.output_tokens": {
             "type": "double",
             "value": 240.0
+          },
+          "gen_ai.cost.reasoning.output_tokens": {
+            "type": "double",
+            "value": 0.0
           },
           "gen_ai.cost.total_tokens": {
             "type": "double",