+++ relay-event-schema/src/protocol/span.rs
@@ -215,7 +215,6 @@ pub struct SentryTags {
     #[metastructure(pii = "true", field = "user.email")]
     pub user_email: Annotated<String>,
     pub environment: Annotated<String>,
-    pub transaction: Annotated<String>,
     #[metastructure(field = "transaction.method")]
     pub transaction_method: Annotated<String>,
     #[metastructure(field = "transaction.op")]
@@ -424,7 +423,6 @@ impl Getter for SentryTags {
             "trace.status" => &self.trace_status,
             "transaction.method" => &self.transaction_method,
             "transaction.op" => &self.transaction_op,
-            "transaction" => &self.transaction,
             "ttfd" => &self.ttfd,
             "ttid" => &self.ttid,
             "user.email" => &self.user_email,
@@ -960,7 +958,7 @@ mod tests {
         let data = r#"{
         "foo": 2,
         "bar": "3",
-        "db.system": "mysql",
+        "db.system.name": "mysql",
         "code.filepath": "task.py",
         "code.lineno": 123,
         "code.function": "fn()",
@@ -1005,7 +1003,7 @@ mod tests {
                 "code.namespace": String(
                     "ns",
                 ),
-                "db.system": String(
+                "db.system.name": String(
                     "mysql",
                 ),
                 "foo": I64(
@@ -1072,7 +1070,7 @@ mod tests {
         );
         assert!(!data.contains("string"));
         assert_eq!(
-            Getter::get_value(&data, "db\\.system"),
+            Getter::get_value(&data, "db\\.system\\.name"),
             Some(Val::String("mysql"))
         );
         assert_eq!(