+++ relay-server/src/processing/errors/errors/unreal.rs
@@ -5,6 +5,8 @@ use crate::managed::{Counted, Quantities, RecordKeeper};
 use crate::processing::ForwardContext;
 use crate::processing::errors::Result;
 use crate::processing::errors::errors::{Context, Expansion, SentryError, utils};
+#[cfg(feature = "processing")]
+use crate::utils::AdditionalExceptions;
 
 #[derive(Debug)]
 pub enum UnrealReport {
@@ -102,14 +104,13 @@ impl SentryError for Unreal {
             if let Some(minidump) = &minidump {
                 crate::utils::process_minidump(
                     event.get_or_insert_with(Default::default),
-                    minidump,
-                    ctx.processing.project_info
+                    minidump, AdditionalExceptions::Delete
                 );
                 metrics.bytes_ingested_event_minidump = (minidump.attachment_body_size() as u64).into();
             }
             if let Some(acr) = &apple_crash_report {
                 crate::utils::process_apple_crash_report(
-                    event.get_or_insert_with(Default::default), ctx.processing.project_info
+                    event.get_or_insert_with(Default::default), AdditionalExceptions::Delete
                 );
                 metrics.bytes_ingested_event_applecrashreport = (acr.len() as u64).into();
             }